refactor: rename eventGridSearchClicked to eventGridSearchTriggered
Renamed function and file from `eventGridSearchClicked` → `eventGridSearchTriggered`
and `event-grid-search-clicked` → `event-grid-search-triggered` for accuracy.
